refactor: wrap pages and templates in layout
