+++ src/pages/404.jsx
@@ -1,13 +1,13 @@
 import React from 'react'
-import SEO from '../components/SEO'
+import {Layout, SEO} from '../components'
 
 function NotFoundPage() {
     return (
-        <>
+        <Layout>
             <SEO title="404: Not Found" />
             <h1>Not Found</h1>
             <p>You just hit a route that doesn&#39;t exist... the sadness.</p>
-        </>
+        </Layout>
     )
 }
